# Add RowPositionsAppender