--- a/core/trino-main/src/main/java/io/trino/operator/output/PositionsAppenderFactory.java
+++ b/core/trino-main/src/main/java/io/trino/operator/output/PositionsAppenderFactory.java
@@ -16,6 +16,7 @@ package io.trino.operator.output;
 import io.trino.spi.block.Int128ArrayBlock;
 import io.trino.spi.block.Int96ArrayBlock;
 import io.trino.spi.type.FixedWidthType;
+import io.trino.spi.type.RowType;
 import io.trino.spi.type.Type;
 import io.trino.spi.type.VariableWidthType;
 import io.trino.type.BlockTypeOperators;
@@ -65,6 +66,9 @@ public class PositionsAppenderFactory
         }
         else if (type instanceof VariableWidthType) {
             return new SlicePositionsAppender(expectedPositions, maxPageSizeInBytes);
+        }
+        else if (type instanceof RowType) {
+            return RowPositionsAppender.createRowAppender(this, (RowType) type, expectedPositions, maxPageSizeInBytes);
         }
 
         return new TypedPositionsAppender(type, expectedPositions);