--- a/core/trino-spi/src/main/java/io/trino/spi/block/AbstractRowBlock.java
+++ b/core/trino-spi/src/main/java/io/trino/spi/block/AbstractRowBlock.java
@@ -51,7 +51,7 @@ public abstract class AbstractRowBlock
     protected abstract boolean[] getRowIsNull();
 
     // the offset in each field block, it can also be viewed as the "entry-based" offset in the RowBlock
-    protected final int getFieldBlockOffset(int position)
+    public final int getFieldBlockOffset(int position)
     {
         int[] offsets = getFieldBlockOffsets();
         return offsets != null ? offsets[position + getOffsetBase()] : position + getOffsetBase();