Optimize partitioned exchange for RowType channels
with batch oriented RowPositionsAppender

<!-- Thank you for submitting a pull request! Find more information in our development guide at https://github.com/trinodb/trino/blob/master/.github/DEVELOPMENT.md and contact us on #dev in Slack. -->

## Description

<!-- Elaborate beyond the title of the PR as necessary to help the reviewers and maintainers.-->

<!-- Answer the following questions to help reviewers and maintainers
understand this PR's scope at a glance:
-->

> Is this change a fix, improvement, new feature, refactoring, or other?

performance improvement
> Is this a change to the core query engine, a connector, client library, or the SPI interfaces? (be specific)

core query engine and spi
> How would you describe this change to a non-technical end user or system administrator?



### Benchmarks
#### tpch/tpcds orc sf1000 partitioned
There is a slight (1.5%) CPU improvement. This is expected as `PartitindOutputOperator` is about 5% of the overall CPU.
![image](https://user-images.githubusercontent.com/8080198/172877960-78830427-f649-4635-9ceb-557f7da049ee.png)

[poo-row-type-sf1000-orc-part.pdf](https://github.com/trinodb/trino/files/8871530/poo-row-type-sf1000-orc-part.pdf)

#### jmh
There is 2x to 4x improvement.
before
```
Benchmark                                   (channelCount)  (enableCompression)  (nullRate)  (partitionCount)  (positionCount)                 (type)  Mode  Cnt    Score    Error  Units
BenchmarkPartitionedOutputOperator.addPage               1                false           0                16             8192      ROW_BIGINT_BIGINT  avgt   20  625.371 ± 12.527  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false           0                16             8192  ROW_RLE_BIGINT_BIGINT  avgt   20  518.245 ± 82.994  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192      ROW_BIGINT_BIGINT  avgt   20  557.170 ± 65.907  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192  ROW_RLE_BIGINT_BIGINT  avgt   20  473.040 ± 64.605  ms/o
```
after
```
Benchmark                                   (channelCount)  (enableCompression)  (nullRate)  (partitionCount)  (positionCount)                 (type)  Mode  Cnt    Score   Error  Units
BenchmarkPartitionedOutputOperator.addPage               1                false           0                16             8192      ROW_BIGINT_BIGINT  avgt   20  144.102 ± 5.855  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false           0                16             8192  ROW_RLE_BIGINT_BIGINT  avgt   20  109.074 ± 1.223  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192      ROW_BIGINT_BIGINT  avgt   20  281.255 ± 4.728  ms/op
BenchmarkPartitionedOutputOperator.addPage               1                false         0.2                16             8192  ROW_RLE_BIGINT_BIGINT  avgt   20  162.054 ± 2.670  ms/op
```

This also brings big improvements for queries with a large number of aggregations that use `RowType` as an intermediate state e.g. `sum`.
This sample query sees about a 30% improvement.
```
trino:tpch_sf1000_dec_orc_part> explain analyze select cast ((orderkey+partkey) %1300000 as int), sum(suppkey), sum(suppkey + 1), sum(suppkey + 2), sum(suppkey + 3),
                             -> sum(suppkey + 4), sum(suppkey + 5), sum(suppkey + 6), sum(suppkey + 7), 
                             -> sum(suppkey + 8), sum(suppkey + 9), sum(suppkey + 10), sum(suppkey + 11)
                             -> from lineitem group by cast ((orderkey + partkey) % 1300000  as int);

```

before
```
Query 20220610_101803_00005_8baqg, FINISHED, 7 nodes
http://localhost:8082/ui/query.html?20220610_101803_00005_8baqg
Splits: 3,081 total, 3,081 done (100.00%)
CPU Time: 32273.5s total,  186K rows/s, 1.88MB/s, 68% active
Per Node: 20.1 parallelism, 3.73M rows/s, 37.6MB/s
Parallelism: 140.5
Peak Memory: 9.59GB
3:50 [6B rows, 59.1GB] [26.1M rows/s, 263MB/s]
```
after
```
Query 20220610_090149_00009_b42pz, FINISHED, 7 nodes
http://localhost:8081/ui/query.html?20220610_090149_00009_b42pz
Splits: 3,081 total, 3,081 done (100.00%)
CPU Time: 22725.7s total,  264K rows/s, 2.66MB/s, 75% active
Per Node: 16.7 parallelism,  4.4M rows/s, 44.4MB/s
Parallelism: 116.7
Peak Memory: 9.87GB
3:15 [6B rows, 59.1GB] [30.8M rows/s, 311MB/s]
```
<!-- The following sections are filled in by the maintainer with input from the contributor:
Use :white_check_mark: or (x) to signal selection.
-->

## Documentation

(x) No documentation is needed.
( ) Sufficient documentation is included in this PR.
( ) Documentation PR is available with #prnumber.
( ) Documentation issue #issuenumber is filed, and can be handled later.

## Release notes

( ) No release notes entries required.
( X) Release notes entries required with the following suggested text:

```markdown
# Section
* Improve performance of queries with a large number of aggregations that use `RowType` as an intermediate state (e.g. `sum`).
```