@lukasz-stec 2022-06-27
comments addressed + commit message extended with the SPI change justification.